+++ examples/ne_A3.jl
@@ -1,37 +1,37 @@
 @testset "test ne A3" begin
 
-    @testset "A3 minlp ne" begin
-         @info "Testing A3 minlp ne"
-         result = run_ne("../examples/data/matgas/A3.m", MINLPGasModel, minlp_solver)
+    @testset "A3 dwp ne" begin
+         @info "Testing A3 dwp ne"
+         result = run_ne("../examples/data/matgas/A3.m", DWPGasModel, minlp_solver)
          @test result["termination_status"] == LOCALLY_SOLVED || result["termination_status"] == OPTIMAL
          @test isapprox(result["objective"], 1781; atol = 1e-1)
      end
 
-    @testset "A3 nlp ne" begin
-        @info "Testing A3 nlp ne"
-        result = run_ne("../examples/data/matgas/A3.m", NLPGasModel, minlp_solver)
+    @testset "A3 wp ne" begin
+        @info "Testing A3 wp ne"
+        result = run_ne("../examples/data/matgas/A3.m", WPGasModel, minlp_solver)
         @test result["termination_status"] == LOCALLY_SOLVED || result["termination_status"] == OPTIMAL
         @test isapprox(result["objective"], 1781; atol = 1e-1)
     end
 
-    @testset "A3 misocp case" begin
-        @info "Testing A3 misocp ne"
-        result = run_ne("../examples/data/matgas/A3.m", MISOCPGasModel, misocp_solver)
+    @testset "A3 crdwp case" begin
+        @info "Testing A3 crdwp ne"
+        result = run_ne("../examples/data/matgas/A3.m", CRDWPGasModel, misocp_solver)
         @test result["termination_status"] == LOCALLY_SOLVED || result["termination_status"] == OPTIMAL
         @test isapprox(result["objective"], 1781; atol = 1e-1)
     end
 
-    @testset "A3 lp case" begin
-        @info "Testing A3 lp ne"
-        result = run_ne("../examples/data/matgas/A3.m", LPGasModel, mip_solver)
+    @testset "A3 lrwp case" begin
+        @info "Testing A3 lrwp ne"
+        result = run_ne("../examples/data/matgas/A3.m", LRWPGasModel, mip_solver)
         @test result["termination_status"] == LOCALLY_SOLVED || result["termination_status"] == OPTIMAL
         @test isapprox(result["objective"], 0.0; atol = 1e-1)
         GC.gc()
     end
 
-    @testset "A3 mip case" begin
-        @info "Testing A3 mip ne"
-        result = run_ne("../examples/data/matgas/A3.m", MIPGasModel, mip_solver)
+    @testset "A3 lrdwp case" begin
+        @info "Testing A3 lrdwp ne"
+        result = run_ne("../examples/data/matgas/A3.m", LRDWPGasModel, mip_solver)
         @test result["termination_status"] == LOCALLY_SOLVED || result["termination_status"] == OPTIMAL
         @test isapprox(result["objective"], 0.0; atol = 1e-1)
     end