+++ src/GasModels.jl
@@ -65,11 +65,11 @@ module GasModels
     include("core/solution.jl")
     include("core/ref.jl")
 
-    include("form/mip.jl")
-    include("form/lp.jl")
-    include("form/nlp.jl")
-    include("form/minlp.jl")
-    include("form/misocp.jl")
+    include("form/lrdwp.jl")
+    include("form/lrwp.jl")
+    include("form/wp.jl")
+    include("form/dwp.jl")
+    include("form/crdwp.jl")
 
     include("prob/gf.jl")
     include("prob/ne.jl")