+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - renamed pd_min, pd_max to be pd_sqr_min and pd_sqr_max (breaking)
 - added pressure slack node constraints to base formulations
 - added compressor energy constraints to base formulations
+- renamed formulations: NLP -> WP (Weymouth Physics), MINLP -> DWP (Disjunctive Weymouth Physics), MISOCP -> CRDWP (Convex relaxation of the disjunctive weymouth physics), MIP -> LRDWP (Linear relaxation of disjunctive weymouth physics), LP -> LRWP (Linear relaxation of the weymouth physics) (breaking)
 
 ## v0.7